+++ components/project/project-detail-tags.tsx
@@ -20,8 +20,10 @@ interface TagsProps extends HtmlHTMLAttributes<HTMLDivElement> {
 
 const TagsWrapper = ({ label, children }: TagsProps) => {
   return (
-    <div className="flex flex-col items-start md:flex-row md:items-center md:gap-2">
-      <span className="py-2 text-base font-medium ">{label}</span>
+    <div className="flex flex-col items-start md:gap-2">
+      <span className="py-2 text-[22px] font-bold text-tuatara-700">
+        {label}
+      </span>
       {children}
     </div>
   )
